Zone Efficiencies
This currently is incomplete. I am not sure what to do when teams never make it to a given zone.

Also reminder that a lot of this is the opposite of how we usually think about it since this is for the defense. If a winner has a high red zone effeciency, that means they often stopped their opponents in the red zone.